x/scichain/keeper/msg_server_profile_test.go
package keeper

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/require"

	"github.com/paranauerj/SciChain/x/scichain/types"
)

func TestProfileMsgServerCreate(t *testing.T) {
	srv, ctx := setupMsgServer(t)
	creator := "A"
	for i := 0; i < 5; i++ {
		resp, err := srv.CreateProfile(ctx, &types.MsgCreateProfile{Creator: creator})
		require.NoError(t, err)
		assert.Equal(t, i, int(resp.Id))
	}
}

func TestProfileMsgServerUpdate(t *testing.T) {
	creator := "A"

	for _, tc := range []struct {
		desc    string
		request *types.MsgUpdateProfile
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgUpdateProfile{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdateProfile{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdateProfile{Creator: creator, Id: 10},
			err:     sdkerrors.ErrKeyNotFound,
		},
	} {
		tc := tc
		t.Run(tc.desc, func(t *testing.T) {
			srv, ctx := setupMsgServer(t)
			_, err := srv.CreateProfile(ctx, &types.MsgCreateProfile{Creator: creator})
			require.NoError(t, err)

			_, err = srv.UpdateProfile(ctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
			}
		})
	}
}

func TestProfileMsgServerDelete(t *testing.T) {
	creator := "A"

	for _, tc := range []struct {
		desc    string
		request *types.MsgDeleteProfile
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgDeleteProfile{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgDeleteProfile{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
		{
			desc:    "KeyNotFound",
			request: &types.MsgDeleteProfile{Creator: creator, Id: 10},
			err:     sdkerrors.ErrKeyNotFound,
		},
	} {
		tc := tc
		t.Run(tc.desc, func(t *testing.T) {
			srv, ctx := setupMsgServer(t)

			_, err := srv.CreateProfile(ctx, &types.MsgCreateProfile{Creator: creator})
			require.NoError(t, err)
			_, err = srv.DeleteProfile(ctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
			}
		})
	}
}
